requirements.txt
fastapi
uvicorn[standard]
kafka-python